archive: contents of `old` moved underneath `snapshot_0`
The contents of `old` are  identical to the current
`unstable/wasi_unstable_preview0` in every sense except the name of the
module and file.

So, for this snapshotting, rather than move the contents of the current
"unstable" (which will be renamed "snapshot" as part of this PR) under a
new directory under phases/old, we're just moving the current contents
under a directory, so that future snapshots can be archived without
changing these paths.