Add functions for working with file permissions.
Add functions and types to query and set the filesystem permissions of
a file: read, write, execute (for files), and search (for directories).

These permissions aren't the only thing which determines whether a given
file or directory can be accessed; hosts may impose additional arbitrary
security restrictions not reflected here.

WASI itself does not currently have the ability to execute files, so the
meaning of "execute" here is entirely up to the host filesystem.

These are similar to fchmod/fchmodat/fstat/fstatat in POSIX, though
there are some differences:

 - This doesn't surface POSIX's setuid, setgid, or sticky bits. These can
   be added in the future if needed, though they'll require additional
   security and portability considerations.

 - This uses separate flags for "execute" on files vs "search" on
   directories. WASI libc can provide a POSIX-compatible C API.

 - This doesn't expose POSIX's user/group/other concepts. These can be
   added in the future if needed, though they'll require additional
   security and portability considerations. Implementations in POSIX
   environments should follow the umask when setting permissions flags.