Add simple way of generating Markdown links in docs
This commit adds a simple way for generating Markdown links in docs
sections of relevant AST nodes. Any slice of the form '`...`' is treated
as a potential link. If the expanded Markdown form (with "::" replaced
with '.') doesn't exist, log a warning message, and leave the link
untouched in its original form.