Change witx unions to be tagged (#220)
* unions are now tagged by an enum

* union fields may be empty. TODO: write tests

* unions: add more tests. compute layout.

* snapshot 0: translate untagged union syntax to tagged union syntax

* snapshot: update from untagged union to tagged union syntax

* snapshot: update docs

* snapshot 0: update docs

* ephemeral: translate untagged unions to tagged unions

* render: fix union output

* bump witx to 0.8.0 - this is definitely semver breaking

* snapshot: I forgot to delete the tag field `subscription.type`

* regen docs

* unionlayout: derive the usual suspects

* snapshot 1 and 0: put tag back in event struct, flatten event_u

prior to this PR, event_u was a union with only one variant
fd_readwrite. the change into tagged union ended up changing the
memory layout of the event struct. so, rather than use tagged
unions in the event struct, we just use a struct that you ignore
in the case of the clock event, which has the same ABI as previously.

in ephemeral, we use a tagged union for events, because we dont have
any ABI stability guarantees to maintain there.

* layout: refactor so we have impls for NamedType and Type too

* ast: change tag of UnionDatatype to be a NamedType

this is much more convenient for consumers, and it reflects the
requirement of the syntax to provide an Id

* fixup docs