Backport fd as handle to past and current snapshots
This commit backports fd as handle to past and current snapshots.
I'm hoping this could prove useful in situations where currently
in the implementor's code, any syscall that deals with `fd`s has
to be marked as potentially "unsafe", whereas with `fd` as a handle
this would be shifted to the fact that `fd` *is* a handle rather
than an alias to an int type.