Add sock_accept()
This function allows a server to accept incoming connections. A few
notes are in order.

Although there are legitimate questions about the capabilities model in
numerous networking APIs, this particular API is both immensely useful
and not a violation of the capabilities model. There is still no way to
create an arbitrary socket. The only thing that WASM can do is accept
incoming connections on a predefined listening socket.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>