Reorganize all of the preview1 specs and tooling.
Rewrite legacy/README.md and legacy/optional-imports.md to reflect the
current status of preview1.

Move the preview0 and preview1 witx specs and docs to legacy/preview0
and legacy/preview1, and remove the ephermal directory. I also backported
the documentation improvements that were made in the ephemeral directory
to the preview1 spec.

Also move the witx tools into legacy/tools.