Remove the `size` type. (#84)
Use `filesize` in `fadvise`, `pread`, and `pwrite`, instead of `size`.
This makes the API independent of wasm32 vs. wasm64 or even linear
memory vs. other options.

This does mean that a wasm32 program can request a `pread` of more bytes
than it can possibly recieve, but it'll just get a short read.