Change `write-repeated` to `write-zeroes`. (#17)
The main use case for `write-repeated` is writing zeros to a bytestream.
I had generalized it to `write-repeated` in anticipation of thinking
about how it will generalize to typed streams in the future, as not all
types have natural zero values. However, not all types can be naturally
copied either, such as owned handles and streams. Consequently, I don't
think we'll generalize this to a fully-general typed-stream function, so
we can specialize it to its intended use case of just writing zeros to
a stream.