Import `datetime` and the stream types from wasi-clocks and wasi-io. (#95)
* Import `datetime` and the stream types from wasi-clocks and wasi-io.

Add `read-via-stream`, `write-via-stream`, and `append-via-stream`
methods to read from, write to, and append to files. And use the
wasi-clocks datetime time for filesystem timestamps.

This uses the same low-tech dependency mechanism as wasi-clocks and
wasi-io are currently using to depend on wasi-poll, which will likely
evolve.

* Fix the return type of `readdir`.

It will eventually be a `stream`, but for now we're using a
`dir-entry-stream` resource as a pseudo-stream.