Add initial wasi-cli content, following the wasi-proposal-repo format.
Add initial wasi-cli content, following the wasi-proposal-repo format.
Define a wit file, add the dependencies in a deps directory, and populate
the README.

For now, this wasi-cli repo defines an additional interface for
exit.wit, which may go away when we can depend on exception-handling,
but is useful for now.

This world is presently missing a few things, notably sockets, but I'll
integrate those once WebAssembly/wasi-sockets#16 lands.