Improve the proposal repo layout. (#7)
* Improve the proposal repo layout.

 - Change from `.wit.md` files with
   wit-embedded-in-markdown-code-comments to just `.wit` files
   containing just Wit. Feedback from early users so far has been that
   people want to just maintain plain wit.

 - Use the new `--html-in-md` option to generate the prettier HTML
   output in the `.md` file, which is rendered on Github. The specifics
   here may evolve, but overall the idea is:
    - Don't check in `.html` files into the repo, as they can't be
      rendered and as such are confusing.
    - Do check in a generated `.md` file, with as much formatting and
      linkifying as we can get.

The generated HTML has a *lot* of room for improvement. But at least with this
PR the HTML is viewable from the Github UI and it has a table of contents and
the formatting is a little tidier.

* Add a `deps` directory example.

Add an example dependency using a `deps` directory. And rename some
things for clarity.

* Update to wit-abi-0.10.0.

This updates to the 0.3.0 release of wit-bindgen.