Rename things to clarify commands, and add a reactor interface.
Renae the "cli" interface and wit file to "command". With the new deps
layout, this means users see cli/command rather than cli/cli.

Accompanying that change, rename the "command" function to "main". This
is not the C/C++ or other language `main` function, but it is the
component-model main function, and at that scope I think calling it
"main" makes sense.

Also, add a "reactor" interface, which is the same as the "command"
interface, except without the "main" function.