descriptor-flags: eliminate non-blocking flag, and therefore set-flags
A blocking or non-blocking read/write is a determined by which method is
used on a read/write stream created from a file. Each of these stream
methods must be implemened for all files. Therefore, it does not make
sense to specify non-blocking at the file level.

The set-flags method was only to be used to set or clear the
non-blocking flag, so this entire method may now be eliminated.