types.wit: delete file mode accessors and setters
We believe there is value in supporting file modes for many applications, but
for the sake of shipping Preview 2, we are out of time to implement these
methods and test a a cross platform implementation of these in two separate
engines.

These functions may come back as part of an additional filesystem interface in
a post Preview 2 patch version 0.2.n, or in Preview 3 or beyond. Until then,
their design is still in the git history.