Revise the documentation for WASI 0.3.0's `tcp-socket::receive`.
`receive` no longer returns its own `result`, so update the text that
says it fails if `receive` has already been called. Instead say that
a new stream that immediately fails is returned.

And, be specific about how the `stream` and the `future` to say that
either both succeed or both fail. This eliminates ambiguity about
whether the `stream` could report a failure independently of the
`future`.