Remove Preview1's advice about random-number generators. (#664)
Previously, the preview1 documentation for the `random_get` function said:

> This function may execute slowly, so when large mounts of random data are
> required, it's advisable to use this function to seed a pseudo-random
> number generator, rather than to provide the random data directly.

However, Wasm guest code is unaware of VM forks in any VMs might be running
in, as well as snapshot/resume features offered by some Wasm engines and
tools, so in practice, WASI use cases have tended toward using `random_get`
in place of guest PRNGs, and thus to relying on `random_get` executing
quickly.

This pattern seems sufficiently widely applicable to motivate updating
WASIp1 to reflect it. To be sure, this is not a behavior change; it's just
removing what we now understand to be bad advice in general.

WASIp2 for its part has already made a similar change.