Initialize a `specifications/` directory (#833)
This initializes a basic structure for our previously published specifications, starting with WASI 0.2.0. For the moment all we do is link to the published WIT files, but in the future this will be helpful to clarify which Wasm and/or Wasm Component Model features are required. As well as include additional prose and spec text present that is not necessarily captured by the WIT interfaces.